Fixed some minor issues with bags and holster
Tapping Y then N quickly while empty handed, with a gun holstered and a
bag on will result in the bag disappearing.

Allowed weapons to be holstered/unholstered while the player is moving
again. Added a Z velocity check to stop players using this to break
falls (which was why the animation check was added at first which also
stopped switching while moving as a side effect)