Vehicle lock update
Added vehicle keys and lock status to the saving code for vehicles.

Added SetVehicleExternalLock which will lock the vehicle and mark it as
locked for restarts. This function isn't used for locking doors while
the player is inside (with KEY_SUBMISSION/2) so that if a player is
inside a vehicle upon restart, they don't get locked out of their
vehicle.

Disabled crowbar lock-breaking for vehicles that are externally locked
with a locksmith kit. I plan to add something in later that allows
players to get into manually locked vehicles.