Tiny changes
DeleteAccount now has an AccountExists check, this will make the command
"/deleteaccount" return the correct error message if a non-existent
username is entered.

For consistency, changed the forward line for the tutorial
OnPlayerCloseContainer hook.

Future plan: Update all ALS hooks to ALS7 (thanks to @Y-Less and
ipsBruno for the clever '#if defined' psychicness!)