Update to defences movement/rotation
Added a different set of rotation fields for horizontal rotation of
defences. Instead of simply adding 90 to the vertical X angle, these
individual rotation values will be used for setting/moving a defence to
the horizontal position.

Tweaked loot spawn rates for gate and panel.

Added the correct holding offsets for the new defence item types.