Improvements to vehicle ownership and saving
When an owned vehicle is saved by another player (not the owner) the
ownership of that vehicle is swapped to the non-owners last vehicle.
This means that entering another players vehicle (perhaps by accident or
just to check fuel levels) will NOT delete the vehicle completely, it
will simply "give" the owner a new vehicle ID. Similarly, if the
intruding player gets back in his own vehicle, the ownership is
transferred BACK to the original state and the vehicle he checked out
won't get deleted upon restart.

Also shifted actual vehicle save code to a separate function to separate
the save code from the ownership transfer code.

Removed some old code, refer to older commits for this now.