Fixed some client crashes
Removed the "Rusty Metal Sheet" item from in-world loot spawns. This
caused client crashes if a rusty sheet spawned inside a single-player
garage with auto-open doors. It was much easier to simply remove the
item from loot spawns than try and find all the loot spawn points that
were in garages.

There is still an issue which needs to be addressed: if a player drops a
rusty sheet inside one of those garages, their game will crash and
anyone who goes near will also crash. (Need a library of all
single-player garages so dropping this specific item in them can be
disabled)