Updated chat colours, anti flood and admin cmds
Moved admin commands to scripts titled by level number for easier
navigation.

Rearranged commands to better suit staff roles:
Game master is for player control but nothing more, they can mute and
warn for chat related offences. They cannot enter duty or spectate. This
position can't be abused to gain an advantage in the game.
Moderator and Administrator are more for catching hackers and have the
tools that can be abused.

Will soon rearrange developer commands because that Level4 file is a
mess.

Updated chat colours for local and radio, these channels show in a
separate colour for distinguishing from global chat easier.

Fixed problems with anti-flood muting, this system now uses the actual
muting function so it will block "/g [text]" properly as well as display
the time remaining until the unmute.