Admin tools and small accounts update
Refined the ban command some more, now it bans the player upon command
usage with empty "reason" and "duration" parameters. After the admin has
typed in a reason and duration in the menus, the ban record is updated.
(In future, this update method will be integrated into the "/banlist"
menu)

Unfortunately, the ban options had to be removed (for now) they may be
reintegrated but will require some new code to extract a player's
inventory/character data for dropping their items to simulate a death as
well as updating their "alive" status.

Added an "/isbanned" command to check if a name is banned.

Added an experimental "/vtp" command.

Added a register check to the CheckForExtraAccounts function. Players
who are new to the server and haven't created an account shouldn't be
checked aliases now due to how the IP is loaded from the database
instead of the actual player's IP (so the function can be used on
offline usernames)

Fixed a bug with the ban update SQLite statement, the ORDER BY
instruction shouldn't have been there.