Merged "SS-Wasteland" world scripts
Seeing as the "SS-Wasteland" modification was only a world related
change, the scripts are now stored here and found in "World_Wasteland".

Not updated though, still quite unfinished but there are plans to
continue this project. For now the files will live here and I may open
up a separate repo in the future that will just contain these world
files instead of a completely separate branch of SS code that misses
updates.