Added a ban warning for aliases lists
Alias functions now check if any of the listed usernames are banned.
Useful for checking ban evaders. Updated the "/aliases" command and the
login alias list to use this.

Alias functions now return the full count of accounts if the loop runs
over the 'max' value. After running over the limit, no more names are
concatenated to the result string but the loop continues and names are
still checked for admin levels and bans.