Minor fixes and updates to food and commands
Moved commands out of the "Admin" folder (finally), renamed the
"Command" script to "CmdProcess" because that's the code that processes
commands rather than a script containing command functions.

Added a staff list variable for the "/admins" command (it used to be
hardcoded with my own server admins). This list is loaded from the
"settings.json" file under the "staff" key.

Fixed a bug with food, if the food item was removed while being eaten
for some reason (given away, deleted, placed on campfire/bbq) the food
could still be eaten.

Fixed a bad debug message in HackDetect.