Vehicle core code update
Note: This file needs to be split up into more logical files.

Removed veh_CurrentModelGroup, no idea why it was in the wrong script
for so long!

Added a delay for saving vehicles, if the player is only in the vehicle
for less than a second then it won't save (this counteracts the problem
of hacking into vehicles and saving them despite being ejected
instantly)

Fixed a major issue with the ResetVehicle function. It now properly
reassigns all the data to the new vehicle ID. The problem was, if a
vehicle had previously been destroyed the the newly created vehicle in
ResetVehicle occupied a different vehicle ID (the lowest free slot).