Improvements and fixes to detection fields
detfield:
- Added GetDetectionFields to return the total number of detection
fields.
- Added GetDetectionFieldLogEntries to return the total number of log
entries for a field.

Fixed a bug where exceptions would "leak" across fields when loading a
field with less exceptions than the previous because the previous ones
would still be in the list as it wasn't emptied on each iteration.

Added DialogPages.pwn which adds simple dialog page buttons.

Added paging for detfield list and detfield log entry list using the new
DialogPages library.

- Known issue: does not look good on smaller resolutions, dialog buttons
need to be moved.