Updates to admin tools and anti combat log code
Added a per-player in-world 3D text tag visible only to admins on duty.
Also indicates whether or not a player is in line-of-sight of the
camera.

Added an extra parameter to "Logout" that can skip the combat log check.

Added combat log check ignore to going on admin-duty so admins won't die
when going on duty after being damaged.

Added checks to ensure taking/giving damage to on-duty admins won't
initiate anti combat log code.