Fixes and improvements to vehicle locking
Fixed a false report that would report players for entering a locked
vehicle if someone already in the vehicle locked it while someone was
entering.

Shifted vehicle lock code to a separate script, provided interface
functions and converted scripts to use them.

Added a validity check to IsPlayerSpawned to avoid array out-of-bounds
errors.

Updated FILE-LIST.md to reflect recent changes.