Shifted nametags to separate script
NameTags.pwn now contains nametag toggle function.
Tags are toggled on when an admin goes on spectate mode, not when they
go on duty to minimise the possibility for abuse (abuse can be checked
using the standard /spec usage checks). Tags are then disabled when the
admin goes off duty (not when they exit spec)

Also added a level 4 command for toggling nametags at any time (even
when not on duty). Can be heavily abused so beware who is level 4.