Completely rewritten player save/load code
Player data now uses modio to store data. Items in the inventory and bag
are stored using ItemLists and ItemArrayData allowing for a variable
amount of data to be stored with items (up to 256 cells).

FileCheck.pwn automatically converts data to modio format, set
"file-check" to 1 in the server.json and open the server once, set it
back to 0 and continue running as normal.