Updated torsos to use item array data
Torsos are the first item to utilise the new array data functionality.
Now, instead of the single cell of extra-data pointing to a separate
"gravestone" entity that contains the name and death reason, the name
and death reason is stored inside the item's array data meaning it will
save with the item.

This means when you kill someone, their name stays with their torso
forever. Very useful for bounties and hired assassins proving they
killed someone.