Fix to alt-tab check, added anti "ninja jack"
"ninja jack" occurs when hijacking a player from the passenger seat and
tapping the sprint or jump key which results in the driver being killed.
This script rectifies that by teleporting the driver out of the vehicle
when the hijacker taps the keys.

A rather clumsy fix and might result in the driver getting frustrated or
thinking the hijacker is a hacker but it's better than players dying for
no valid reason.