Item duplication exploit fix to backpacks
Multi-tapping the add-to-inventory key while wearing a backpack and
having a full inventory would cause the same item to be added to the
container multiple times. If the player reconnects, they are recreated
as individual and fully usable items.

This also warrants a fix in SIF that disallows the same item ID to be
added to a container (and inventory) multiple times.