Added improved hash and IP logging scripts
These scripts will log a player's GPCI or IP when the join the server
and it is different to the existing record. This serves two purposes: a
history of a player's IP and GPCI values and as a way of
cross-referencing accounts to find accounts owned by the same person or
accessed from the same location. Soon will be a DNS/ISP log too.