Defence script overhaul (+modio implementation)
Defences are now saved and loaded using modio.

The rotation is dictated by the "pose" field which also has a new type
for a future update.

Defences now need a motor in order to have a keypad installed. Defences
with just a motor can be opened and closed without entering a code.

If a motor is present, the "pose" value is ignored and "moveState" is
used instead (this dictates the open/closed state of the defence)