Improved defence keypads
The regular keypad on a defence now uses the keypad UI instead of
dialogs to prevent the use of external brute-force applications to break
codes.

y_inline doesn't appear to work so for now, it uses standard callbacks.

Added an Advanced Keypad which, when installed on a defence, takes a 4-8
character hexadecimal string (stored as a 32 bit integer) meaning there
are a lot more combinations available for passcodes.

Some minor fixes here and there in defences and keypad code.

Improved some UI feedback for opening/closing defences.