Fixed some issues with the whitelist and settings
OnGameModeInit has been further split into "Setup" and "Pre" (all of
this will be cleaned up a bit more in the future with a cleaner sequence
of callbacks and better hooks.)

Whitelist settings moved to whitelist script.

Auto-whitelist now uses OnPlayerLogin for checks.

Added pre definition that increases the container size which will
prevent an OOB error in SIF.