Fix to vehicle area overflow
Fixed a problem where the list of nearby vehicles would not be
size-checked before attempting to append to.

Also fixed a minor issue with handling return values from
OnPlayerInteractVehicle so a return of 1 will (as intended) break the
looping of nearby vehicles. This bug resulted in multiple trunks being
opened at once.