Added note item
Note item can store up to 256 characters of text. Blank notes are
writable and notes with text on can be torn up but not edited. Added to
all loot spawn indexes as a common item and added a few default messages
with the option to expand via the settings.json file.