Added plant pots and seeds with functionality
Plant pots can now have seeds placed in them. Seeds need water to grow
which is added with a Bottle (Water). Seed types are defined with a
"growth time" which is measured in server restarts. A growth time of 3
would require the server to restart 3 times to grow fully (example:
starting at 0 | 1 | 2 | 3, grown where | is a restart).

Currently only one seed type that produces a Tomato item which is
currently useless.

This update is dependent on a SIF change that is coming soon. This
change allows allocation of item IDs before actually creating the item.
This is used in Tent.pwn (and will be used elsewhere) to ensure item
array data is set BEFORE the item is created and thus before
OnItemCreate/InWorld is called which may want to access the array data.

Updated bottle item with some minor code improvements. A larger update
to bottles is planned which will come when the entire food system is
reviewed.