Updated aliases, join logs and playerlist module
The aliases functions now return a simple list of players rather than a
data structure with names and banned states. Banned state is now checked
after the list is acquired.

"/aliases" command now displays a PlayerList dialog instead of a basic
menu, this greatly improves the amount of interaction options available
for players in the list for advanced searching and analysis.

Improved ipv4-log and gpci-log to allow searching the selected item in
the list with the opposing method (an item in an IP lookup can be
searched by name or an item in a name lookup can be searched by IP).
Warning: currently a memory issue with the GPCI history lookup, may be
an issue with sqlitei.

Added highlightbanned optional parameter to ShowPlayerList.

Fixed some issues with ShowPlayerList.