Minor fixes to items, account API and bags
When inventory and bag items are loaded, items are created as virtual
items to avoid unnecessary calls to OnItemCreateInWorld as well as
accidentally removing the items from bag containers due to some code in
SIF that automatically removes an item from its container upon being
created in the world. The callbacks are usually called in an order that
doesn't result in this but sometimes (usually on high server load) the
OnItemCreateInWorld was called after the item was added to the container
resulting in it being removed instantly.

Fixed a typo in the GetAccountAliveState API function.

Fixed a typo in backpacks that would get the free slots of the wrong
container resulting in items vanishing in an attempt to move them to a
container with not enough free slots.

Added back the 'exit' RCON call on OnGameModeExit, this causes the
server to crash when restarting but is unfortunately the only way to
address a problem with sockets on Linux.