Overhaul to backpacks and inventory code
(Warning: depends on changes made in a recent SIF update)

Restricted carry items in backpacks and inventory.

Resized a lot of large and carry items so they are smaller but still
won't fit in backpacks or inventories.

Unified inventory and container item add checks with the same technique
with new SIF return values for AddItemToInventory/Container.

Found and fixed some small bugs in the process, mostly involving items
disappearing when trying to move them to a full inventory or container.