Overhauled car-armour (carmour) module
Revived the Carmour module (finally!) with support for vehicletypes. A
carmour type is bound to a vehicletype with SetVehicleTypeCarmour which
takes a name as a parameter. This name corresponds to a file (without
the path or extension) in the directory
"./scriptfiles/vehicles/carmour/" which should contain nothing more than
CreateObject lines which create a bunch of objects around the null point
(0,0,0) which acts as the vehicle centre.

The three unique carmour vehicle data files are not packaged with the
server since they don't belong to me so you'll get three
SetVehicleTypeCarmour errors when starting the server, just ignore these
or create the required files with your own creations!

Added Moonbeam due to popular demand.

Added Duneride for shits and gigs.