Updated weapon calculator application
Added shots to kill and time to kill outputs.

Added knockout checkbox that will become checked if the simulation
results in a knockout based on the knockout chance which gets higher
depending on the knockout multiplier, wound count and total bleed rate.

Updated the unconsciousness chance label in the inventory screen.