Fixed /vsave
Fixed the problem with the /vsave command. When you crashed or reloged
the server did not save the option you set with /vsave.
It was annoying, because if you set /vsave to off, and you left your
vehicle at base then you crash far away from base the /vsave set itself
to on, after you join. So you couldn't save the vehicle at your base, if
you entered another one.

I am using MySQL R39 for saving, and I haven't used SQLite, so there
might be mistakes in the code. And also i could not test it, because the
compiler crashes for me.