Improved loading functions for permanent entities
WARNING: Requires the latest modio revision which as of the time of this
commit, isn't currently available. Until that's released, just use the
commit before this one.

Improved the validity checks and modio calls for all 4 main permanent
entities (vehicles, defences, safeboxes and tents). modio files are now
closed properly when needed and proper validity checks are performed
when file-reads are attempted.