Removed JSON for good, replaced with SimpleINI
SimpleINI has had no commits for over a year but at least it works.
Replaced DJson since it was dumping an empty .db file in the
/scriptfiles/ directory for EVERY restart for unknown reasons. It also
didn't load most of the string arrays or object arrays.

For now, removed the ability to customise the respawn/newspawn items.

Removed preset text on notes.

Increased defence limit.