New lock item, boat tweaks, spectate fix
Added "Lock and Chain" item that can be used to lock bikes and boats (or
any vehicle that can't be locked using a locksmith kit).

Fixed the interaction zone for the Marquis boat.

Changed the threshold distance for boat movement sync resets.

Fixed a bug with spectate UI displaying knockout time wrong.