Updates to knockouts and vehicle deletion
Partly reverted the changes to vehicle deletion. This time, if a vehicle
is removed via the world (explosion, sinking and the ever-annoying
random desync deletion) it will simply respawn. This will be refined
once a better way to detect how a vehicle was destroyed is implemented.
Added back the full deletion functionality to DestroyWorldVehicle.

Added logging to knockouts and wakeups.

Added check to weapon damage that ensures knocked out players can't mod
animation changes or take advantage of desync to damage players while
they should be knocked out.