Added loot spawns to backpacks (again)
Partially fixed again. It can cause a recursion problem if the loot
spawn amount is too high.

Changed the behaviour of item creation in FillContainerWithLoot so
OnItemCreate is called after the item is added to the container so
GetItemContainer actually returns a value. This uses the rarely touched
preallocation feature of SIF/Item!