Adjusted map loading sequence to give feedback
Addresses issue #162

This change adds a timer between each district loading function which
allows the server to communicate with players for a short time (100ms)
so they don't lose connection. Each district loading function also sends
a client message to all players to let them know that the server is
actually doing something (in time, regular players will learn the
sequence and judge the progress for themselves.)

This can still fail if the server can't process the connection cookie in
the 100ms window between districts but this appears to rarely happen and
most of the time a client will wait in the "cookie lobby" until the
current district has finished and opens up the server for 100ms.

Also changed the mapname from a global variable to a function.

Removed supply crate debug mode, it appears to be working perfectly now.

Added some information about copying and credits.